(chore) move SessionLog to sub-subfolder
